# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: HyperTide Backend API
  version: 2.0.0-draft
  description: HyperTide v2 API draft contract for M8 baseline freeze.
servers:
  - url: http://localhost:3000
paths:
  /health/live:
    get:
      summary: Process liveness
      responses:
        '200': { description: Live }
  /health/ready:
    get:
      summary: Dependency readiness
      responses:
        '200': { description: Ready }
        '503': { description: Not ready }
  /metrics:
    get:
      summary: Prometheus-compatible operational metrics
      responses:
        '200':
          description: Metrics in Prometheus text exposition format
          content:
            text/plain:
              schema:
                type: string
  /v2/auth/verify:
    get:
      summary: Verify API key validity
      security:
        - ApiKeyAuth: []
      responses:
        '200': { description: Verified }
  /v2/auth/generate:
    post:
      summary: Generate API key
      security:
        - ApiKeyAuth: []
      responses:
        '201': { description: Created }
  /v2/auth/revoke:
    delete:
      summary: Revoke API key
      security:
        - ApiKeyAuth: []
      responses:
        '200': { description: Revoked }
  /v2/auth/keys:
    get:
      summary: List API keys
      security:
        - ApiKeyAuth: []
      responses:
        '200': { description: Listed }
  /v2/auth/exchange-key:
    post:
      summary: Exchange API key for JWT tokens
      responses:
        '200': { description: Token pair issued }
  /v2/auth/refresh:
    post:
      summary: Refresh access token
      responses:
        '200': { description: Refreshed }
  /v2/auth/revoke-refresh:
    post:
      summary: Revoke refresh token
      responses:
        '200': { description: Revoked }
  /v2/locks/acquire:
    post:
      summary: Acquire file lock
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Locked }
  /v2/locks/release:
    post:
      summary: Release file lock
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Unlocked }
  /v2/locks/renew:
    post:
      summary: Renew file lock lease
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Renewed }
  /v2/locks/force-release:
    post:
      summary: Force release lock
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Released }
  /v2/locks:
    get:
      summary: List locks
      parameters:
        - in: query
          name: repo_id
          required: true
          schema: { type: string }
        - in: query
          name: scope
          required: false
          schema: { type: string, default: asset }
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Listed }
  /v2/storage/upload:
    post:
      summary: Upload blob
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Uploaded }
  /v2/storage/download/{hash}:
    get:
      summary: Download blob
      parameters:
        - in: path
          name: hash
          required: true
          schema: { type: string }
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Downloaded }
  /v2/storage/exists/{hash}:
    get:
      summary: Check blob exists
      parameters:
        - in: path
          name: hash
          required: true
          schema: { type: string }
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Exists result }
  /v2/storage/hash:
    post:
      summary: Calculate hash
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Hashed }
  /v2/blobs/missing:
    post:
      summary: Query missing chunk hashes
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Missing chunk list }
  /v2/blobs/chunks/{chunk_hash}:
    put:
      summary: Upload chunk by hash (idempotent)
      parameters:
        - in: path
          name: chunk_hash
          required: true
          schema: { type: string }
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Uploaded or already exists }
  /v2/manifests:
    post:
      summary: Create canonical manifest and persist metadata
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '201': { description: Created }
  /v2/repos:
    get:
      summary: List repositories
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Repository list }
    post:
      summary: Create repository
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repo_id]
              properties:
                repo_id: { type: string }
                default_branch: { type: string, default: main }
      responses:
        '201': { description: Created }
        '400': { description: Invalid repo id or default branch }
        '409': { description: Repository already exists }
  /v2/repos/{repo_id}:
    get:
      summary: Get repository details
      parameters:
        - in: path
          name: repo_id
          required: true
          schema: { type: string }
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Repository details }
        '404': { description: Repository not found }
  /v2/branches:
    post:
      summary: Create branch
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '201': { description: Created }
  /v2/branches/{repo_id}:
    get:
      summary: List branches
      parameters:
        - in: path
          name: repo_id
          required: true
          schema: { type: string }
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Listed }
  /v2/changesets:
    post:
      summary: Submit changeset
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '201': { description: Created }
  /v2/changesets/{changeset_id}/approve:
    post:
      summary: Approve draft changeset
      parameters:
        - in: path
          name: changeset_id
          required: true
          schema: { type: string }
        - in: query
          name: repo_id
          required: true
          schema: { type: string }
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Approved }
  /v2/changesets/{changeset_id}/promote:
    post:
      summary: Promote approved changeset to visible head
      parameters:
        - in: path
          name: changeset_id
          required: true
          schema: { type: string }
        - in: query
          name: repo_id
          required: true
          schema: { type: string }
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Promoted }
  /v2/changesets/{changeset_id}/gate:
    get:
      summary: Query approve gate status for promote
      parameters:
        - in: path
          name: changeset_id
          required: true
          schema: { type: string }
        - in: query
          name: repo_id
          required: true
          schema: { type: string }
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Gate status }
  /v2/history/{repo_id}:
    get:
      summary: List history
      parameters:
        - in: path
          name: repo_id
          required: true
          schema: { type: string }
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Listed }
  /v2/rollback:
    post:
      summary: Submit rollback changeset
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '201': { description: Created }
  /v2/sync/{repo_id}:
    get:
      summary: Sync snapshot
      parameters:
        - in: path
          name: repo_id
          required: true
          schema: { type: string }
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Synced }
  /v2/trust/checkpoints/generate:
    post:
      summary: Generate trust checkpoint
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '201': { description: Checkpoint generated }
  /v2/trust/checkpoints/latest:
    get:
      summary: Get latest trust checkpoint
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Latest checkpoint }
        '404': { description: Checkpoint not found }
  /v2/trust/checkpoints/{checkpoint_id}/witness/attest:
    post:
      summary: Add witness receipt for checkpoint
      parameters:
        - in: path
          name: checkpoint_id
          required: true
          schema: { type: string }
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Witness receipt recorded }
  /v2/trust/witness/summary:
    get:
      summary: Get witness quorum summary
      parameters:
        - in: query
          name: checkpoint_id
          required: true
          schema: { type: string }
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Witness summary }
  /v2/trust/witness/topology:
    get:
      summary: Get witness topology configuration with environment groups
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Witness topology including cross-environment quorum metadata }
  /v2/trust/audit/verify:
    post:
      summary: Verify audit hash-chain integrity
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Verify result }
  /v2/trust/audit/export:
    get:
      summary: Export audit chain entries (compliance)
      parameters:
        - in: query
          name: limit
          required: false
          schema: { type: integer }
        - in: query
          name: before_seq
          required: false
          schema: { type: integer }
        - in: query
          name: action
          required: false
          schema: { type: string }
        - in: query
          name: actor_id
          required: false
          schema: { type: string }
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Audit entries exported }
  /v2/trust/retention/policy:
    get:
      summary: Get retention policy configuration
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Retention policy }
  /v2/trust/replay/verify:
    post:
      summary: Verify event replay consistency against current state
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Replay verification result }
  /v2/trust/replay/readiness:
    get:
      summary: Evaluate event-led readiness based on replay consistency
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200': { description: Replay readiness report }
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        success: { type: boolean, enum: [true] }
        data:
          nullable: true
        error:
          nullable: true
    ErrorEnvelope:
      type: object
      properties:
        success: { type: boolean, enum: [false] }
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              nullable: true
            request_id: { type: string }